+++ frontend/src/api/playlist-api-service.ts
@@ -1,15 +1,18 @@
 import ApiService, {DefaultApiService} from "./api-service";
 import {Observable, throwError} from "rxjs";
 import {PlaylistRequest} from "../model/playlist-request";
-import {PlaylistResponse} from "../model/playlist";
+import {PlaylistItem, PlaylistResponse} from "../model/playlist";
 
 const PLAYLIST_API_PATH = 'playlist';
 const TARGET_UPDATE_API_PATH = 'playlist/update';
+const REVERSE_URL_API_PATH = 'playlist/reverse';
 
 export default interface PlaylistApiService extends ApiService {
     getPlaylist(req: PlaylistRequest): Observable<PlaylistResponse>;
 
     updateTargets(targets: string[]): Observable<any>;
+
+    getReverseUrl(item: PlaylistItem): Observable<string>;
 }
 
 export class DefaultPlaylistApiService extends DefaultApiService implements PlaylistApiService {
@@ -23,5 +26,9 @@ export class DefaultPlaylistApiService extends DefaultApiService implements Play
 
     updateTargets(targets: string[]): Observable<any> {
         return this.post(TARGET_UPDATE_API_PATH, targets);
+    }
+
+    getReverseUrl(item: PlaylistItem): Observable<string> {
+        return this.post(REVERSE_URL_API_PATH, item);
     }
 }