+++ src/api/model/config.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 
 use crate::model::api_proxy::ApiProxyConfig;
-use crate::model::config::{LogConfig, ProcessingOrder, ReverseProxyConfig, ScheduleConfig, WebAuthConfig};
+use crate::model::config::{LogConfig, ProcessingOrder, ReverseProxyConfig, ScheduleConfig, WebUiConfig};
 use crate::model::config::{ConfigApi, ConfigRename, ConfigSort, ConfigTargetOptions, InputType, MessagingConfig, TargetOutput, VideoConfig};
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
@@ -53,8 +53,6 @@ pub struct ServerConfig {
     pub api_proxy: Option<ApiProxyConfig>,
     pub log: Option<LogConfig>,
     pub update_on_boot: bool,
-    pub web_ui_enabled: bool,
-    pub web_auth: Option<WebAuthConfig>,
-
+    pub web_ui: Option<WebUiConfig>,
 }